requirements.txt
asgiref
boto3
botocore
Django
django-cors-headers
django-storages
djangorestframework
gunicorn
jmespath
Pillow
psycopg2-binary
pytz
s3transfer
six
sqlparse
toml
urllib3
protobuf==3.18.1
grpcio
h5py
idna
pytz
termcolor
numpy
geopy
geographiclib
Werkzeug
wrapt
tensorflow-cpu
keras